Refactored the RangeTest, added Jdoc.
Changed the RangeTest by splitting it intodifferent tests that helps understand the way a range it's created and used.
