+++ stdlib/Foundation/Bound.java
@@ -9,7 +9,18 @@
  */
 package Foundation;
 
+/**
+ * Instance of Bound.java
+ * 
+ * @author 
+ * @version
+ */
 public interface Bound<T extends Comparable<T>> extends Comparable<Bound<T>> {
-	
+
+	/**
+	 * Returns the value of the bound.
+	 * 
+	 * @return the value of the bound.
+	 */
 	public T value();
 }
