+++ test/ci/AllTests.java
@@ -1,0 +1,26 @@
+/*
+ * This source file is part of the swift open source project.
+ *
+ * Copyright (c) 2017 Guillermo Facundo Colunga and the swift project authors.
+ * Licensed under GNU General Public License v3.0.
+ *
+ * See /LICENSE for license information.
+ * 
+ */
+package ci;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+import Foundation.RangeTest;
+
+/**
+ * Instance of AllTests.java
+ * 
+ * @author 
+ * @version 
+ */
+@RunWith(Suite.class) @SuiteClasses({ RangeTest.class }) public class AllTests {
+
+}
